bpf/nfsops.c: defend against possible stale pointer
When we trap the kretprobe, we lookup the start_t pointer
to take from it the information like inode, byte-count and
start-time. Make sure to only delete it from the map _after_
we finished dereferencing it to prevent a theoretical case
where this struct becomes stale.

Note that I am not sure if this is even theoretically possible,
but better be safe than sorry. There is practically no overhead
delaying the map delete for a few more instructions.

Signed-off-by: Sagi Grimberg <sagi@grimberg.me>