nfsops: execute modprobe and xz on host machine using combination of exec + chroot
To avoid using init container for modprobe and mouting additional host folders for xz exection
I implemented approch that allows host based execution in container enviroment.

In docker file there are symlinks created for command that needed to be used on host.
Each time when such command is executed in container chroot is happening and temporary move execution to host.