nfsops.py: try to get pid envs from stats collection path
If  EnvTracer did not catch the pid envs in time the pid was sampled
from ebpf or it was dropped due to some load on the perf event ring,
we will try to opportunistically search for the envs from the collection
path.

If we are trying to get_pid_envs from the sampling path, we allow empty {}
entries in the pid map to avoid repeating scans.

Signed-off-by: Sagi Grimberg <sagi@grimberg.me>