nfsops.py: make mounts map refresh to simply update and not filled from scratch
We simply update the mounts map with new information, and once every
vaccum_interval we clear it. Note that if the map is very stable, and
we don't find new pids that have private mounts, it will not be
vaccumed, but that is very much fine, because vaccuming is only
intended to prevent uncontrolled growth of the map.

Signed-off-by: Sagi Grimberg <sagi@grimberg.me>