vdb: keep vdb sessions short lived
Open a new session for each sample, and then tear it down.
The main reason is that we want as much spreading of collectors
to different cnodes as possible.

This will not be very efficient for static vips, but for a DNS name
which is the intended use, we want that each sample will resolve to
a different cnode.

Signed-off-by: Sagi Grimberg <sagi@grimberg.me>