main: speedup signal handler
Because the signal handler is caught outside of the event loop, it is
not seen by the event loop correctly.

This causes systemctl stop to wait until <interval> is met (which may
be long minutes at times). Fix that by setting the stop_event from the
event loop, which makes the signal handler effect immediate,
and in turn also service stop.

Signed-off-by: Sagi Grimberg <sagi@grimberg.me>